Move file transfer code to separate module
